requirements.txt
strands-agents==0.1.7
strands-agents-tools
httpx
wikipedia
pydantic
python-dotenv
